src/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
batch-face==1.5.4
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
keras==3.13.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-transforms==0.0.6
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
Pygments==2.19.2
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sixdrepnet==0.1.6
soupsieve==2.8.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
Werkzeug==3.1.4
wheel==0.45.1
wrapt==2.0.1
